WIP: Add deletionProtection and autoStart support for VirtualMachine

+++ internal/controller/proxmox/virtualmachine_controller.go
@@ -153,21 +153,12 @@ func (r *VirtualMachineReconciler) Reconcile(ctx context.Context, req ctrl.Reque
 		}
 		metrics.IncVirtualMachineCount()
 	} else {
-		// If exists, check if it is running or not
-		// If not running, start the VM
-		vmState := proxmox.GetVMState(vmName, nodeName)
-		if vmState == "stopped" {
-			proxmox.StartVM(vmName, nodeName)
-		} else {
-			// Update the VirtualMachine spec if needed
-			proxmox.UpdateVM(vmName, nodeName, vm)
-			err = r.Update(context.Background(), vm)
-			if err != nil {
-				logger.Error(err, "Error updating VirtualMachine")
-				return ctrl.Result{}, client.IgnoreNotFound(err)
-			} else {
-				logger.Info(fmt.Sprintf("VirtualMachine %s updated", vmName))
-			}
+		// Check if auto start is enabled
+		r.handleAutoStart(ctx, vm)
+		err = r.UpdateVirtualMachine(ctx, vm)
+		if err != nil {
+			logger.Error(err, "Error updating VirtualMachine")
+			return ctrl.Result{Requeue: true}, client.IgnoreNotFound(err)
 		}
 	}
 	logger.Info(fmt.Sprintf("VirtualMachine %s already exists and running", vmName))
@@ -234,8 +225,13 @@ func (r *VirtualMachineReconciler) CreateVirtualMachine(vm *proxmoxv1alpha1.Virt
 func (r *VirtualMachineReconciler) DeleteVirtualMachine(vm *proxmoxv1alpha1.VirtualMachine) {
 	// Delete the VM
 	kubernetes.CreateVMKubernetesEvent(vm, kubernetes.Clientset, "Deleting")
-	proxmox.DeleteVM(vm.Spec.Name, vm.Spec.NodeName)
-	metrics.DecVirtualMachineCount()
+	if vm.Spec.DeletionProtection {
+		metrics.DecVirtualMachineCount()
+		return
+	} else {
+		proxmox.DeleteVM(vm.Spec.Name, vm.Spec.NodeName)
+		metrics.DecVirtualMachineCount()
+	}
 }
 
 func (r *VirtualMachineReconciler) UpdateVirtualMachineStatus(vm *proxmoxv1alpha1.VirtualMachine) error {
@@ -264,5 +260,34 @@ func (r *VirtualMachineReconciler) handleResourceNotFound(ctx context.Context, e
 		return nil
 	}
 	logger.Error(err, "Failed to get VirtualMachine")
+	return err
+}
+
+func (r *VirtualMachineReconciler) handleAutoStart(ctx context.Context,
+	vm *proxmoxv1alpha1.VirtualMachine) ctrl.Result {
+	logger := log.FromContext(ctx)
+	if vm.Spec.EnableAutoStart {
+		vmName := vm.Spec.Name
+		nodeName := vm.Spec.NodeName
+		vmState := proxmox.GetVMState(vmName, nodeName)
+		if vmState == "stopped" {
+			proxmox.StartVM(vmName, nodeName)
+			logger.Info(fmt.Sprintf("VirtualMachine %s is started", vmName))
+			return ctrl.Result{Requeue: true}
+		}
+	}
+	return ctrl.Result{}
+}
+
+func (r *VirtualMachineReconciler) UpdateVirtualMachine(ctx context.Context, vm *proxmoxv1alpha1.VirtualMachine) error {
+	logger := log.FromContext(ctx)
+	updateStatus := proxmox.UpdateVM(vm)
+	err := r.UpdateVirtualMachineStatus(vm)
+	if err != nil {
+		return err
+	}
+	if updateStatus {
+		logger.Info(fmt.Sprintf("VirtualMachine %s is updated", vm.Spec.Name))
+	}
 	return err
 }
